% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSIPrefix.R
\docType{methods}
\name{BestSIPrefix-methods}
\alias{BestSIPrefix-methods}
\alias{BestSIPrefix,EPhysData-method}
\alias{BestSIPrefix,EPhysData,EPhysSet,ANY-method}
\alias{BestSIPrefix,EPhysSet-method}
\alias{SetSIPrefix}
\alias{SetSIPrefix,EPhysData-method}
\alias{SetSIPrefix,EPhysData,EPhysSet,ANY-method}
\alias{SetSIPrefix,EPhysSet-method}
\title{Set the SI prefix for an EPhysData or EPhysSet object}
\usage{
BestSIPrefix(X)

\S4method{BestSIPrefix}{EPhysData}(X)

\S4method{BestSIPrefix}{EPhysSet}(X)

SetSIPrefix(X, value)

\S4method{SetSIPrefix}{EPhysData}(X, value)

\S4method{SetSIPrefix}{EPhysSet}(X, value)
}
\arguments{
\item{X}{An EPhysData or EPhysSet object}
}
\value{
The EPhysData or EPhysSet object with units converted to the new SI prefix.
}
\description{
Sets the SI prefix to a user-define value or so that it minimizes the number of relevant digits for a vector of units.
}
\examples{
## BestSIPrefix
X<-makeExampleEPhysData()
X@Data<-X@Data/1000
ggEPhysData(X)
X<-BestSIPrefix(X)
ggEPhysData(X)

 X<-makeExampleEPhysSet()
 X<-BestSIPrefix(X)

 ## SetSIPrefix
X<-makeExampleEPhysData()
ggEPhysData(X)
X<-SetSIPrefix(X,"k")
ggEPhysData(X)

 X<-makeExampleEPhysSet()
 X<-SetSIPrefix(X,"k")
}
