% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Data_ChannelMetadata-methods.R
\name{ChannelMetadata}
\alias{ChannelMetadata}
\alias{AddChannelMetadata}
\alias{ChannelMetadata<-}
\title{Channel Metadata Accessors for EPhysContainer}
\usage{
ChannelMetadata(X, column = NULL, ...)

AddChannelMetadata(X, columnName, columnData, ...)

ChannelMetadata(X) <- value
}
\value{
\describe{
\item{\code{ChannelMetadata(X)}}{The channel metadata data.frame.}
\item{\code{ChannelMetadata(X, column="...")}}{A vector ordered by \code{Channels(X)}.}
\item{\code{AddChannelMetadata(X, columnName, columnData)}}{Modified object.}
\item{\code{ChannelMetadata(X) <- value}}{Modified object.}
}
}
\description{
Channel Metadata Accessors for EPhysContainer
}
\details{
Get or modify the \code{Channel_Metadata} slot of an \code{EPhysContainer}.
Optionally fetch a single column by name (returned as a vector aligned to \code{Channels}).
}
