% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Data_dim.R
\docType{methods}
\name{dim}
\alias{dim}
\alias{dim,EPhysData-method}
\alias{dim,EPhysContinuous-method}
\title{Dimensions for EPhys* objects}
\usage{
\S4method{dim}{EPhysData}(x)

\S4method{dim}{EPhysContinuous}(x)
}
\arguments{
\item{x}{An \link{EPhysData} or \link{EPhysContinuous} object.}
}
\value{
An integer vector of length 2: number of rows and columns in \code{x@Data}.
}
\description{
Returns the matrix-like dimensions of the \verb{@Data} slot.
}
\details{
Supported classes:
\itemize{
\item \link{EPhysData}
\item \link{EPhysContinuous}
}
}
\section{Functions}{
\itemize{
\item \code{dim(EPhysData)}: Dimensions of an \link{EPhysData} object.

\item \code{dim(EPhysContinuous)}: Dimensions of an \link{EPhysContinuous} object.

}}
\examples{
\dontrun{
dim(ephys_data)        # for EPhysData
dim(ephys_continuous)  # for EPhysContinuous
}

}
