% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim,EPhysData-method}
\alias{dim,EPhysData-method}
\title{Dimensions of an EPhysData object}
\usage{
\S4method{dim}{EPhysData}(x)
}
\arguments{
\item{x}{An object of class \link{EPhysData}.}
}
\value{
A vector of length 2 containing the number of rows and columns in the data slot.
}
\description{
This method returns the dimensions of the data slot in an \link{EPhysData} object.
}
