% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downsample.R
\docType{methods}
\name{Downsample}
\alias{Downsample}
\alias{Downsample,EPhysData-method}
\alias{Downsample,EPhysSet-method}
\title{Downsample EPhysData or EPhysSet Object}
\usage{
Downsample(X, n = 250)

\S4method{Downsample}{EPhysData}(X, n = 250)

\S4method{Downsample}{EPhysSet}(X, n = 250)
}
\arguments{
\item{X}{An \link{EPhysData} or \link{EPhysContainer} instance, depending on the method.}

\item{n}{integer giving the \emph{desired} number of
    intervals.  Non-integer values are rounded down.}
}
\value{
An object of class \link{EPhysData} or \link{EPhysSet} after downsampling
}
\description{
This method downsamples an \link{EPhysData} or an \link{EPhysSet} object by sampling the time trace and applying a filter function.
}
\details{
The \code{Downsample} function samples the time trace of the input object and applies a filter function to the data. The sampling is based on the \link[base:pretty]{bsae::pretty} function, which creates a sequence of about equally spaced values that cover the range of the input time trace. The filter function stored in the object is applied before downsampling and gets relaced  by an identity function.
}
\examples{
# For an EPhysData object
myEPhysData <- makeExampleEPhysData(sample_points=500, replicate_count = 3)
myEPhysData
downsampled_myEPhysData <- Downsample(myEPhysData)
downsampled_myEPhysData

# For an EPhysSet object
myEPhysSet <- makeExampleEPhysSet(sample_points=750, nsets=10)
myEPhysSet
myEPhysSet[[1]]
downsampled_myEPhysSet <- Downsample(myEPhysSet)
downsampled_myEPhysSet
downsampled_myEPhysSet[[1]]

}
