% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPhysData.R
\docType{package}
\name{EPhysData-package}
\alias{EPhysData-package}
\title{EPhysData: Structured Index and Introduction for the EPhysData R Package}
\description{
There are two classes for storing the data. The \code{EPhysData} class stores  data from a single recording (possibly with repeated measurements). It is usually generated from imported raw data using \code{newEPhysSet()}. Information on how to filter or average the data contained in this class can be stored in it, but the data is left unchanged. The \code{EPhysSet} class contains a collection of associated  \code{EPhysData} classes (e.g. data from different channels recorded in parallel or in response to increasing stimulus intensities). Each \code{EPhysData} stored is associated with metadata describing the nature of the particular recodring. This class can be subsetted, extended or modified in other ways.
}
\section{Object creation}{

\itemize{
\item \link[=newEPhysData]{newEPhysData} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=newEPhysSet]{newEPhysSet} (for \link[=EPhysSet]{EPhysSet} objects) \cr\cr
}
}

\section{Object manipulation}{

\itemize{
\item \link[=AverageFunction<-]{AverageFunction<-} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=FilterFunction<-]{FilterFunction<-} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=Rejected<-]{Rejected<-} (for \link[=EPhysData]{EPhysData} objects) \cr\cr
\item \link[=AddData]{AddData<-} (for \link[=EPhysSet]{EPhysSet} objects) \cr
\item \link[=Metadata]{Metadata<-} (for \link[=EPhysSet]{EPhysSet} objects) \cr
\item \link[=AddMetadata]{AddMetadata} (for \link[=EPhysSet]{EPhysSet} objects) \cr
\item \link[=lapply]{lapply} (for \link[=EPhysSet]{EPhysSet} objects) \cr\cr
}
}

\section{Accession functions}{

\itemize{
\item \link[=Subset]{Subset} (for \link[=EPhysData]{EPhysData} and \link[=EPhysData]{EPhysSet}  objects) \cr
\item \link[=[]{[} (for \link[=EPhysData]{EPhysData}) \cr
\item \link[=[[]{[[} and \link[=[[<-]{[[<-} (for \link[=EPhysSet]{EPhysSet}) \cr
\item \link[=as.data.frame]{as.data.frame} (for \link[=EPhysData]{EPhysSet} and \link[=EPhysData]{EPhysSet}  objects) \cr\cr
\item \link[=GetData]{GetData} (for \link[=EPhysData]{EPhysData} objects) \cr\cr
\item \link[=AverageFunction]{AverageFunction} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=FilterFunction]{FilterFunction} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=StimulusTrace]{StimulusTrace} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=TimeTrace]{TimeTrace} (for \link[=EPhysData]{EPhysData} objects) \cr
\item \link[=Rejected]{Rejected} (for \link[=EPhysData]{EPhysData} objects) \cr\cr
\item \link[=Metadata]{Metadata} (for \link[=EPhysSet]{EPhysSet} objects) \cr\cr
}
}

\section{Helper functions}{

\itemize{
\item \link[=makeExampleEPhysData]{makeExampleEPhysData} \cr
\item \link[=makeExampleEPhysSet]{makeExampleEPhysSet} \cr
\item \link[=length]{length} (for \link[=EPhysSet]{EPhysSet} objects) \cr
\item \link[=dim]{dim} (for \link[=EPhysData]{EPhysData} objects) \cr\cr
\item \link[=Save]{Save} Saves EPhysData or EPhySet objects to an HDF5 file \cr
\item \link[=Load]{Load} Loads EPhysData or EPhySet objects from an HDF5 file \cr
}
}

\examples{
# Load the EphysData package
library(EPhysData)

# Create an example EPhysData object
data <- makeExampleEPhysData()

# Get data
GetData(data, Raw = TRUE)

# Get time trace data
TimeTrace(data)

# Set Averaging and function for the data
FilterFunction(data)<-function(x){x/max(x)}
GetData(data, Raw = FALSE)
AverageFunction(data)<-mean
GetData(data, Raw = FALSE)

# Create an EPhysSet object
ephysSet <-
  newEPhysSet(
    Data = list(
      makeExampleEPhysData(),
      makeExampleEPhysData(),
      makeExampleEPhysData(),
      makeExampleEPhysData()
    ),
    Metadata = data.frame(StepID = c("A1", "A2", "A3", "A4"))
    )
ephysSet


}
\author{
\href{https://www.lindnerlab.de}{Moritz Lindner}
}
