% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Class_EPhysIEI-Class.R
\docType{class}
\name{EPhysIEI-class}
\alias{EPhysIEI-class}
\title{EPhysIEI-class}
\description{
Subclass of \code{EPhysContainer} for inter‐event‐interval data. Each
\code{Data[[i]]} is a named list of channels, each a two‐column data.frame
with \code{Time} and \code{IEI}.
}
\keyword{classes}
