% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0EPhysSet-class.R
\docType{class}
\name{EPhysSet}
\alias{EPhysSet}
\title{EPhysSet Class}
\description{
A class representing a set of Electrophysiological recordings (e.g. a time series or data from the same subject obtained in response to different stimuly) together with defining metadata.
}
\section{Slots}{

\describe{
\item{\code{Data}}{Data A list of \linkS4class{EPhysData} objects.}

\item{\code{Metadata}}{A data frame containing metadata information associated with the data, each row corresponds to one list item. Rows must be unique}
}}

