% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Set_EPhysData.R
\docType{methods}
\name{Get_Set_EPhysData}
\alias{Get_Set_EPhysData}
\alias{Rejected,EPhysData-method}
\alias{Rejected<-,EPhysData-method}
\alias{FilterFunction,EPhysData-method}
\alias{FilterFunction<-,EPhysData-method}
\alias{AverageFunction,EPhysData-method}
\alias{AverageFunction<-,EPhysData-method}
\alias{TimeTrace,EPhysData-method}
\alias{StimulusTrace,EPhysData-method}
\title{Get/Set methods for EPhysData objects}
\usage{
\S4method{Rejected}{EPhysData}(X, return.fx = F)

\S4method{Rejected}{EPhysData}(X) <- value

\S4method{FilterFunction}{EPhysData}(X)

\S4method{FilterFunction}{EPhysData}(X) <- value

\S4method{AverageFunction}{EPhysData}(X)

\S4method{AverageFunction}{EPhysData}(X) <- value

\S4method{TimeTrace}{EPhysData}(X)

\S4method{StimulusTrace}{EPhysData}(X)
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{return.fx}{For \code{Rejected()}: Whether to return the function or the resulting logical vector. Default is \code{FALSE}, i.e. to return the function.}

\item{value}{A value (usually a function) to set.}
}
\description{
These methods are used to get and set the non-data slots from \link{EPhysData} objects.
}
\details{
\code{Rejected}: These functions set or get a function returning a logical vector indicating which of the repeated measurements stored in an \link{EPhysData} object to exclude from averaging. The following function from the EPhysMethods package may be helpful: \link[EPhysMethods:autoreject.by.distance]{EPhysMethods::autoreject.by.distance}, \link[EPhysMethods:autoreject.by.signalfree]{EPhysMethods::autoreject.by.signalfree}

\code{FilterFunction}: Set  a function for filtering each individual of the repeated measurements in the \link{EPhysData} object. Could be downsampling or noise removal, for instance. The following functions from the EPhysMethods package may be helpful: \link[EPhysMethods:filter.bandpass]{EPhysMethods::filter.bandpass}, \link[EPhysMethods:filter.detrend]{EPhysMethods::filter.detrend},

\code{AverageFunction}: Set a function describing how averaging across repeated measurement should be performed in the \link{EPhysData} object. Usually, \link[base:mean]{mean} can be a good start.

\code{TimeTrace}: These functions set or get the time trace belonging to the measurements stored in an \link{EPhysData} object.
}
\section{Functions}{
\itemize{
\item \code{Rejected(EPhysData)}: Rejected

\item \code{Rejected(EPhysData) <- value}: Rejected<-

\item \code{FilterFunction(EPhysData)}: FilterFunction

\item \code{FilterFunction(EPhysData) <- value}: FilterFunction<-

\item \code{AverageFunction(EPhysData)}: AverageFunction

\item \code{AverageFunction(EPhysData) <- value}: AverageFunction<-

\item \code{TimeTrace(EPhysData)}: TimeTrace

\item \code{StimulusTrace(EPhysData)}: StimulusTrace

}}
\examples{
# Create an EPhysData object with example data
myEPhysData <- makeExampleEPhysData()

# Get the "Rejected" slot
Rejected(myEPhysData)

# Set the "Rejected" slot
Rejected(myEPhysData) <- c(FALSE, FALSE, TRUE)

# Get the "filter.fx" slot
FilterFunction(myEPhysData)

# Set the "filter.fx" slot
FilterFunction(myEPhysData) <- function(x) x^2

# Get the "average.fx" slot
AverageFunction(myEPhysData)

# Set the "average.fx" slot
AverageFunction(myEPhysData) <- median

}
\seealso{
\link[EPhysMethods:autoreject.by.distance]{EPhysMethods::autoreject.by.distance}, \link[EPhysMethods:autoreject.by.signalfree]{EPhysMethods::autoreject.by.signalfree}, \link[EPhysMethods:filter.bandpass]{EPhysMethods::filter.bandpass}, \link[EPhysMethods:filter.detrend]{EPhysMethods::filter.detrend},
}
