% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEPhysData.R
\docType{methods}
\name{ggEPhysData}
\alias{ggEPhysData}
\alias{ggEPhysData,EPhysData-method}
\alias{ggEPhysData,EPhysData,ANY-method}
\title{Generate a ggplot2 plot for EPhysData objects}
\usage{
ggEPhysData(X, Raw = T, ShowFiltered = T, SetSIPrefix = "keep")

\S4method{ggEPhysData}{EPhysData}(X, Raw = T, ShowFiltered = T, SetSIPrefix = "keep")
}
\arguments{
\item{X}{An \link{EPhysData} or \link{EPhysContainer} instance, depending on the method.}

\item{Raw}{Logical indicating whether to get raw data or processed (filtered - see: \link{FilterFunction}, averaged - see: \link{AverageFunction}) data.}

\item{ShowFiltered}{Apply the filter to the raw traces before showing them.}

\item{SetSIPrefix}{Change the SI prefix. Set to \code{keep} (default), for not to change anything, to \code{auto} for using the \link{BestSIPrefix-methods} to minimize the number of relevant digits or to any SI prefix to use that. Calls the \link{SetSIPrefix}method}
}
\value{
A ggplot2 plot visualizing the EPhysData.
}
\description{
This method generates a ggplot2 plot for objects of class 'EPhysData'. It allows you to visualize electrical physiology data over time.
}
\examples{
ephys_data <- makeExampleEPhysData(,replicate_count = 15)  # Create an EPhysData object
Rejected(ephys_data) <- c(rep(c(FALSE,FALSE,TRUE),5)) # randomly reject some
AverageFunction(ephys_data) <- mean
ggEPhysData(ephys_data)  # Generate a ggplot2 plot
ggEPhysData(ephys_data,SetSIPrefix="k")  # Generate a ggplot2 plot
}
