% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply,EPhysSet,ANY-method}
\alias{lapply,EPhysSet,ANY-method}
\title{Apply a Function to Elements of an EPhysSet Object}
\usage{
\S4method{lapply}{EPhysSet,ANY}(X, FUN, ReturnEPhysSet = T)
}
\arguments{
\item{X}{An \code{EPhysSet} object.}

\item{FUN}{The function to apply to the elements of \code{X@Data}.}

\item{ReturnEPhysSet}{Logical. If \code{TRUE}, the results are returned as a modified \code{EPhysSet} object with the function applied to the data. If \code{FALSE}, the results are returned as a list.}
}
\value{
If ReturnEPhysSet is \code{TRUE} modified \code{EPhysSet} object with the function applied to the data. If ReturnEPhysSet is \code{FALSE}, a list with the results will be returned.
}
\description{
Applies a specified function to elements of an \code{EPhysSet} object. The function is applied to the data contained within the \code{EPhysSet} and the results are returned as a modified \code{EPhysSet} object or as a matrix, depending on the chosen options.
}
\examples{
myEPhysSet <- makeExampleEPhysSet()
myEPhysSet
myEPhysSet<-lapply(myEPhysSet, function(x){FilterFunction(x)<-sin; return(x)})
myEPhysSet
FilterFunction(myEPhysSet@Data[[2]])

}
