% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length,EPhysSet-method}
\alias{length,EPhysSet-method}
\title{Length of an EPhysSet object}
\usage{
\S4method{length}{EPhysSet}(x)
}
\arguments{
\item{x}{An object of class \link{EPhysSet}.}
}
\value{
An integer representing the number of recordings stored in the object.
}
\description{
This method returns the length of the data slot in an \link{EPhysSet} object.
}
