% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSave.R
\docType{methods}
\name{LoadSave}
\alias{LoadSave}
\alias{Save}
\alias{Save,EPhysData-method}
\alias{Save,EPhysData,EPhysSet,ANY-method}
\alias{Save,EPhysSet-method}
\alias{Load}
\title{Load/Save EPhysData or EPhysSet objects to or from  HDF5 files}
\usage{
Save(X, filename, overwrite = F)

\S4method{Save}{EPhysData}(X, filename, overwrite = F)

\S4method{Save}{EPhysSet}(X, filename, overwrite = F)

Load(filename)
}
\arguments{
\item{X}{A \linkS4class{EPhysData} or \linkS4class{EPhysSet} object.}

\item{filename}{Path the data is read from or written to.}

\item{overwrite}{Should existing files be overwritten?}
}
\value{
\itemize{
\item Save: Does not return any values.
\item Load: An \linkS4class{EPhysData} or \linkS4class{EPhysSet} object.
}
}
\description{
These functions load and save \linkS4class{EPhysData} or \linkS4class{EPhysSet} objects from or into HDF5 files.
}
\section{Functions}{
\itemize{
\item \code{Save()}: Save \linkS4class{EPhysData} or \linkS4class{EPhysSet} objects to an HDF5 file

\item \code{Load()}: Load \linkS4class{EPhysData} or \linkS4class{EPhysSet} objects from an HDF5 file

}}
\examples{
fn<-tempfile()
SampleData<-makeExampleEPhysData()
Save(SampleData, fn, overwrite = TRUE)
SampleData
rm(SampleData)
SampleData<-Load(fn)
SampleData
SampleData<-makeExampleEPhysSet()
Save(SampleData, fn, overwrite = TRUE)
SampleData
rm(SampleData)
SampleData<-Load(fn)
SampleData
}
\seealso{
\link[=Load]{Load}, \link[=EPhysData]{EPhysData}, \link[=EPhysSet]{EPhysSet}
}
