% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExample.R
\name{makeExample}
\alias{makeExample}
\alias{makeExampleEPhysData}
\alias{makeExampleEPhysSet}
\title{Create Example EPhysData Object}
\usage{
makeExampleEPhysData(
  sample_points = sample(seq(1, 400, 10), 1),
  start_time = 0,
  time_unit = "ms",
  data_unit = "mV",
  replicate_count = sample(3:7, 1)
)

makeExampleEPhysSet(
  nsets = 4,
  sample_points = sample(seq(1, 400, 10), 1),
  start_time = 1,
  time_unit = "s",
  data_unit = "mV",
  replicate_count = sample(2:6, 1),
  metadata_data = NULL
)
}
\arguments{
\item{sample_points}{Number of rows for the example data frame.}

\item{start_time}{The starting time for the time trace.}

\item{time_unit}{The time unit for the time trace.}

\item{data_unit}{The unit for the data values.}

\item{replicate_count}{The number of replicate columns in the data frame.}

\item{nsets}{Number of data sets to include into an \code{EPhysSet} object.}

\item{metadata_data}{The metadata data frame to be associated with the \code{EPhysSet}.}
}
\value{
An \code{EPhysData} object with example data.
}
\description{
These functions generate an example \code{EPhysData} or \code{EPhysSet} object for demonstration purposes.
}
\section{Functions}{
\itemize{
\item \code{makeExampleEPhysSet()}: makeExampleEPhysSet

}}
