% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Data_Metadata.R
\docType{methods}
\name{Metadata}
\alias{Metadata}
\alias{Metadata<-}
\alias{AddMetadata}
\alias{Metadata,EPhysSet-method}
\alias{Metadata<-,EPhysSet-method}
\alias{AddMetadata,EPhysSet-method}
\alias{Metadata,EPhysContainer-method}
\alias{Metadata<-,EPhysContainer-method}
\alias{AddMetadata,EPhysContainer-method}
\title{Get or modify Metadata on EPhysContainer objects}
\usage{
Metadata(X)

Metadata(X) <- value

AddMetadata(X, columnName, columnData)

\S4method{Metadata}{EPhysSet}(X)

\S4method{Metadata}{EPhysSet}(X) <- value

\S4method{AddMetadata}{EPhysSet}(X, columnName, columnData)

\S4method{Metadata}{EPhysContainer}(X)

\S4method{Metadata}{EPhysContainer}(X) <- value

\S4method{AddMetadata}{EPhysContainer}(X, columnName, columnData)
}
\arguments{
\item{X}{An \link{EPhysContainer} object (or subclass such as \link{EPhysSet}).}

\item{value}{New metadata data.frame (row count must match).}

\item{columnName}{Name of column to add/replace.}

\item{columnData}{Vector to insert.}
}
\value{
For \code{Metadata()}: the metadata \code{data.frame}.
For \code{Metadata<-} and \code{AddMetadata}: the updated object.
}
\description{
Retrieve or modify the \code{Metadata} slot on \link{EPhysSet} and
\link{EPhysContainer}. Use:
\itemize{
\item \code{Metadata(X)} to get the data.frame,
\item \code{Metadata(X) <- value} to replace it entirely,
\item \code{AddMetadata(X, name, data)} to add/replace a single column.
}
}
\section{Functions}{
\itemize{
\item \code{Metadata(EPhysSet)}: Get metadata (EPhysSet)

\item \code{Metadata(EPhysSet) <- value}: Replace metadata (EPhysSet)

\item \code{AddMetadata(EPhysSet)}: Add/replace one column (EPhysSet)

\item \code{Metadata(EPhysContainer)}: Get metadata (EPhysContainer)

\item \code{Metadata(EPhysContainer) <- value}: Replace metadata (EPhysContainer)

\item \code{AddMetadata(EPhysContainer)}: Add/replace one column (EPhysContainer)

}}
\examples{
# Create example data
Data <- data.frame(rep1 = c(0.5, 0.52, 0.48, 0.49),
                   rep2 = c(0.55, 0.57, 0.54, 0.56),
                   rep3 = c(0.48, 0.47, 0.46, 0.48))
TimeTrace <- c(1, 2, 3, 4)
Unit <- "mV"
TimeUnit <- "s"
data1 <- newEPhysData(Data = Data, TimeTrace = TimeTrace, Unit = Unit, TimeUnit = TimeUnit)

# Create metadata and an EPhysSet object
metadata <- data.frame(StepID = c("A1", "A2", "A3", "A4"))
ephysSet <- new("EPhysSet", Data = list(data1, data1, data1, data1), Metadata = metadata)

# Add a new column to the metadata
ephysSet <- AddMetadata(ephysSet, "Condition", c("Control", "Treatment", "Control", "Treatment"))

# Replace the metadata
newMetadata <- data.frame(StepID = c("B1", "B2", "B3", "B4"))
Metadata(ephysSet) <- newMetadata

# Retrieve metadata
retrievedMetadata <- Metadata(ephysSet)
retrievedMetadata

}
