% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{Metadata}
\alias{Metadata}
\alias{AddMetadata}
\alias{Metadata<-}
\title{Get or modify the metadata slot}
\usage{
Metadata(X)

AddMetadata(X, columnName, columnData)

Metadata(X) <- value
}
\arguments{
\item{X}{An \code{EPhysSet} object.}

\item{columnName}{Name of the new column to be added.}

\item{columnData}{Data for the new column.}

\item{value}{New metadata to replace the existing metadata.}
}
\value{
Either an updated \code{EPhysSet} object or it's metadata.
}
\description{
These functions are used to modify (\code{AddMetadata}, \code{Metadata<-}) metadata of or retrieve metadata from an \code{EPhysSet} object.
}
\section{Functions}{
\itemize{
\item \code{AddMetadata()}: AddMetadata

\item \code{Metadata(X) <- value}: Metadata<-

}}
\examples{
# Create example data
Data <- data.frame(rep1 = c(0.5, 0.52, 0.48, 0.49),
                   rep2 = c(0.55, 0.57, 0.54, 0.56),
                   rep3 = c(0.48, 0.47, 0.46, 0.48))
TimeTrace <- c(1, 2, 3, 4)
Unit <- "mV"
TimeUnit <- "s"
data1 <- newEPhysData(Data = Data, TimeTrace = TimeTrace, Unit = Unit, TimeUnit = TimeUnit)

# Create metadata and an EPhysSet object
metadata <- data.frame(StepID = c("A1", "A2", "A3", "A4"))
ephysSet <- new("EPhysSet", Data = list(data1, data1, data1, data1), Metadata = metadata)

# Add a new column to the metadata
ephysSet <- AddMetadata(ephysSet, "Condition", c("Control", "Treatment", "Control", "Treatment"))

# Replace the metadata
newMetadata <- data.frame(StepID = c("B1", "B2", "B3", "B4"))
Metadata(ephysSet) <- newMetadata

# Retrieve metadata
retrievedMetadata <- Metadata(ephysSet)
retrievedMetadata
}
