% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newEPhysData.R
\name{newEPhysData}
\alias{newEPhysData}
\title{Store raw electrophysiological data into a \link{EPhysData} object.}
\usage{
newEPhysData(
  Data,
  TimeTrace,
  StimulusTrace = as_units(integer(0), unitless),
  Unit = NULL,
  TimeUnit = NULL,
  StimulusUnit = NULL
)
}
\arguments{
\item{Data}{A matrix or data.frame containing the raw electrophysiological data.
Columns represent repeated measurements.}

\item{TimeTrace}{A numeric vector representing the time trace of the data.}

\item{StimulusTrace}{A numeric vector representing the stimulus trace (optional).}

\item{Unit}{A character string representing the unit of the electrophysiological data.
It should be a basic or derived SI unit. If the unit has a prefix, it must be converted
into a factor and stored in the 'Factor' slot of the output object. If 'Data' inherits the type 'units', this parameter is optional.}

\item{TimeUnit}{A character string representing the unit of time for the time trace.
It should be one of 's', 'min', 'h', 'd', or 'w'. If 'Data' inherits the type 'units', this parameter is optional.}

\item{StimulusUnit}{A character string representing the unit of the stimulus data (optional).
It should be a basic or derived SI unit. If the unit has a prefix, it must be converted
into a factor and stored in the 'StimulusFactor' slot of the output object. If 'Data' inherits the type 'units', this parameter is optional.}
}
\value{
An object of class \link{EPhysData} containing the raw electrophysiological data and other information.
}
\description{
This function takes in raw electrophysiological stores them into a \link{EPhysData} object.
}
\examples{
Data <- data.frame(rep1 = c(0.5, 0.52, 0.48, 0.49),
                   rep2 = c(0.55, 0.57, 0.54, 0.56),
                   rep3 = c(0.48, 0.47, 0.46, 0.48))
TimeTrace <- c(1, 2, 3, 4)
Unit <- "mV"
TimeUnit <- "s"
newEPhysData(Data = Data, TimeTrace = TimeTrace, Unit = Unit, TimeUnit = TimeUnit)
# Output: An object of class 'EPhysData' with processed data and information.
}
\seealso{
\link{EPhysData}
}
