% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset.R
\name{Subset-method}
\alias{Subset-method}
\alias{Subset,EPhysData-method}
\alias{Subset,EPhysSet-method}
\alias{[[,EPhysSet-method}
\alias{[,EPhysData-method}
\title{Subset from EPhysData Object}
\usage{
Subset(X, ...)

\S4method{Subset}{EPhysData}(
  X,
  Time = range(TimeTrace(X)),
  TimeExclusive = FALSE,
  Repeats = !Rejected(X),
  Raw = T,
  ...
)

\S4method{Subset}{EPhysSet}(
  X,
  Time = NULL,
  TimeExclusive = FALSE,
  Repeats = NULL,
  SetItems = rep(TRUE, nrow(Metadata(X))),
  Raw = T,
  Simplify = F,
  ...
)

\S4method{[[}{EPhysSet}(x, i)

\S4method{[}{EPhysData}(x, i, j)
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{...}{currently unused.}

\item{Time}{Numeric vector of length 2 representing the time range for data extraction.
Default is the entire time range (i.e., keep all data).}

\item{TimeExclusive}{Keep only the two time points stated under 'Time', not the range.}

\item{Repeats}{If X is an EPhysSet, this parameter can only be used if all EPhysData contained in the set has the same number of repeats.
Numeric index/indices or a logical vector of the same length as repeats stored,}

\item{Raw}{Logical indicating whether to get raw data or processed (filtered, averaged) data.}

\item{SetItems}{Which items of the set to subset/keep.}

\item{Simplify}{Logical if 'True' will return \code{EPhysData} instead of \code{EPhysSet} if only one \code{EPhysData} is left in the set.}

\item{i, j}{Indices specifying elements to extract.}
}
\value{
\code{Subset}: An \code{EPhysData} or an \code{EPhysSet} object representing the subsetted data.
}
\description{
This method subsets an \code{EPhysData} or an \code{EPhysSet} object into a new object of the same class.
}
\details{
The \code{Subset} function creates a new \code{EPhysData} or \code{EPhysSet} object containing a subset of the data (and metadata for \code{EPhysSet}) from the original object, based on the provided parameters.
}
\section{Functions}{
\itemize{
\item \code{Subset(EPhysData)}: Subset method for EPhysData

\item \code{Subset(EPhysSet)}: Subset method for EPhysSet

\item \code{[[}: Extract specific items from an \linkS4class{EPhysSet} object. Returns an \linkS4class{EPhysData} object or a list thereof.

\item \code{[}: Extract specific items from an \linkS4class{EPhysData} object. Returns a data.frame.

}}
\examples{
# Subset EPhysData
myEPhysData <- makeExampleEPhysData()

## Get subsetted data based on time range and repeated measurements
subsetted_myEPhysData <- Subset(myEPhysData, Time = TimeTrace(myEPhysData)[c(1, 3)], Repeats = c(1, 2))
subsetted_myEPhysData

# Subset EPhysSet
myEPhysSet <- makeExampleEPhysSet(nsets=10)
subsetted_myEPhysSet <- Subset(myEPhysSet, SetItems=c(4:7))
subsetted_myEPhysSet
Metadata(subsetted_myEPhysSet)
}
\concept{EPhysData-methods}
\concept{Subsetting_Dataextraction}
