% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subset.R
\name{Subset}
\alias{Subset}
\alias{Subset,EPhysData-method}
\alias{Subset,EPhysSet-method}
\alias{make_metadata_parilist}
\title{Subset from EPhysData Object}
\usage{
Subset(X, ...)

\S4method{Subset}{EPhysData}(
  X,
  Time = range(TimeTrace(X)),
  TimeExclusive = FALSE,
  Repeats = !Rejected(X),
  Raw = T
)

\S4method{Subset}{EPhysSet}(
  X,
  Time = NULL,
  TimeExclusive = FALSE,
  Repeats = NULL,
  Metadata_select = make_metadata_parilist(X),
  Raw = T,
  Simplify = F
)

make_metadata_parilist(X, except = list())
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{...}{currently unused.}

\item{Time}{Numeric vector of length 2 representing the time range for data extraction.
Default is the entire time range (i.e., keep all data).}

\item{TimeExclusive}{Keep only the two time points stated under \code{Time}, not the range.}

\item{Repeats}{Specifies which columns of \code{X@Data} to use for extraction.
It can be either a numeric vector specifying the column indices of the repeated measurements
or a logical vector of the same length as columns in \code{X@Data},
where \code{TRUE} indicates using that column for extraction. Default is the inverse of the \code{\code{\link{Rejected}}(X)} vector.}

\item{Raw}{Logical indicating whether to get raw data or processed (filtered, averaged) data.}

\item{Metadata_select}{A pairlist of metadata parameters for subsetting (only for \code{EPhysSet}).}

\item{Simplify}{Logical if 'True' will return \code{EPhysData} instead of \code{EPhysSet} if only one \code{EPhysData} is left in the set.}

\item{except}{Which values to remove in the subset.}
}
\value{
\code{Subset}: An \code{EPhysData} or an \code{EPhysSet} object representing the subsetted data.

\code{make_metadata_parilist}: A pairlist providing input for the 'Metadata_select' parameter of \code{Subset} for \code{EPhysSet}.
}
\description{
This method subsets an \code{EPhysData} or an \code{EPhysSet} object into a new object of the same class.
}
\details{
The \code{Subset} function creates a new \code{EPhysData} or \code{EPhysSet} object containing a subset of the data (and metadata for \code{EPhysSet}) from the original object, based on the provided parameters.

The \code{make_metadata_parilist} function is a helper function for \code{Subset}. It creates a pairlist that provides input for the 'Metadata_select' parameter of \code{Subset} for \code{EPhysSet}. as 'Metadata_select' is a positive selector, \code{make_metadata_parilist} may be particularly helpful when only a single condition should be excluded. See example below.
}
\section{Functions}{
\itemize{
\item \code{Subset,EPhysData-method}: Subset method for EPhysData

\item \code{Subset,EPhysSet-method}: Subset method for EPhysSet

\item \code{make_metadata_parilist}: make_metadata_parilist
}}

\examples{
# Subset EPhysData
myEPhysData <- makeExampleEPhysData()

## Get subsetted data based on time range and repeated measurements
subsetted_myEPhysData <- Subset(myEPhysData, Time = TimeTrace(myEPhysData)[c(1, 3)], Repeats = c(1, 2))
subsetted_myEPhysData

# Subset EPhysSet with make_metadata_parilist
myEPhysSet <- makeExampleEPhysSet()
selection_list <- make_metadata_parilist(myEPhysSet,except=list(StepID=c("A3")))
selection_list
subsetted_myEPhysSet <- Subset(myEPhysSet,Metadata_select=selection_list)
subsetted_myEPhysSet
Metadata(subsetted_myEPhysSet)
}
\concept{EPhysData-methods}
\concept{Subsetting_Dataextraction}
