% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddData.R
\docType{methods}
\name{AddData}
\alias{AddData}
\alias{AddData,EPhysSet-method}
\alias{AddData,EPhysSet,ANY-method}
\title{Add Data to an EPhysSet Object}
\usage{
AddData(X, newData, newMetadata)

\S4method{AddData}{EPhysSet}(X, newData, newMetadata)
}
\arguments{
\item{X}{An \code{EPhysSet} object.}

\item{newData}{An \linkS4class{EPhysData} object to be added to the \code{Data} slot.}

\item{newMetadata}{A data frame containing metadata information associated with the new data.}
}
\value{
An updated \code{EPhysSet} object with the new data and metadata added.
}
\description{
Adds a new \link{EPhysData} object to the \code{Data} slot of an \code{EPhysSet} object and the corresponding metadata to the \code{Metadata} slot.
}
\examples{
# Create an example EPhysSet object
ephys_set <- makeExampleEPhysSet()

# Create new EPhysData and metadata
new_data <- makeExampleEPhysData()
new_metadata <- data.frame(StepID = "B3")

# Add new data to the EPhysSet
updated_ephys_set <- AddData(ephys_set, newData = new_data, newMetadata = new_metadata)
updated_ephys_set
}
