% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,EPhysData-method}
\alias{as.data.frame,EPhysSet-method}
\title{as.data.frame for EPhysData and EPhysSet}
\usage{
\S4method{as.data.frame}{EPhysData}(x, Raw = T, IncludeRejected = F, ...)

\S4method{as.data.frame}{EPhysSet}(x, Raw = T)
}
\arguments{
\item{x}{An \link{EPhysData} or \link{EPhysSet} object.}

\item{Raw}{Logical indicating whether to get raw data or processed (filtered - see: \link{FilterFunction}, averaged - see: \link{AverageFunction}) data.}

\item{IncludeRejected}{Logical. If \code{TRUE}, rejected trials will be included; if \code{FALSE} (default), rejected trials will be dropped.}

\item{...}{currently unused.}
}
\value{
A data frame representing the \link{EPhysData} object in long format. Note that, by default, rejected Trials are excluded.
}
\description{
Converts an \link{EPhysData} or \link{EPhysSet} object to a data frame in a long format with columns for trial number, time, stimulus (if avaliable), and value.
}
\section{Functions}{
\itemize{
\item \code{as.data.frame(EPhysData)}: Method for EPhysData

\item \code{as.data.frame(EPhysSet)}: Method for EPhysSet

}}
\examples{
# Create an example EPhysData object
new_data <- makeExampleEPhysData()
# Convert EPhysData to a data frame
new_data <- as.data.frame(new_data)
head(new_data)

# Create an example EPhysSet object
new_data <- makeExampleEPhysSet()
new_data <- as.data.frame(new_data)
head(new_data)

}
