% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0EPhysData-class.R
\docType{class}
\name{EPhysData-class}
\alias{EPhysData-class}
\alias{EPhysData}
\title{EPhysData - Class for Electrophysiological Data}
\value{
An object of class \code{EPhysData}
}
\description{
This class represents the most fundamental unit of electrophysiological data acquisition, typically obtained in response to a one single type of stimulus. It includes repeated measurements, time and stimulus traces, units, as well as information on post-processing steps (such as filtering and averaging) that can be applied to the data.
}
\section{Slots}{

\describe{
\item{\code{Data}}{A 2D array representing the  electrophysiological data with columns for repeated measurements}

\item{\code{TimeTrace}}{A numeric vector containing the time trace of the electrophysiological data.}

\item{\code{StimulusTrace}}{A numeric vector containing the stimulus trace (optional).}

\item{\code{Created}}{A POSIXct object representing the creation date and time of the object.}

\item{\code{Rejected}}{A logical vector indicating which repeated measurements to exclude from averaging.}

\item{\code{filter.fx}}{A function that can be applied to each repeated measurement By default, it leaves the data unchanged.}

\item{\code{average.fx}}{A function describing how averaging across repeated measurement should be performed. By default, repeated measurements are returned and no averaging is performed.}
}}

\seealso{
\link{newEPhysData}
}
