% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\docType{methods}
\name{GetData}
\alias{GetData}
\alias{GetData,EPhysData-method}
\alias{GetData,EPhysData,ANY-method}
\title{Get Data from EPhysData Object}
\usage{
GetData(
  X,
  Time = range(TimeTrace(X)),
  TimeExclusive = F,
  Trials = NULL,
  Raw = F
)

\S4method{GetData}{EPhysData}(
  X,
  Time = range(TimeTrace(X)),
  TimeExclusive = F,
  Trials = NULL,
  Raw = F
)
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{Time}{Numeric vector of length 2 representing the time range for data extraction.
Default is the entire time range (i.e., keep all data).}

\item{TimeExclusive}{Keep only the two points stated under 'Time', not the range.}

\item{Trials}{Specifies which of the repeated measurements (if any) to use for extraction.
It can be either a numeric vector specifying the indices of the repeated measurements,
a logical vector of the same length as trials stored, where \code{TRUE} indicates using that column for extraction, or
NULL, indicating that the function stored in the Rejected (\link{Rejected}) slot will be used (default).}

\item{Raw}{Logical indicating whether to get raw data or processed (filtered - see: \link{FilterFunction}, averaged - see: \link{AverageFunction}) data.}
}
\value{
A data matrix containing either raw or processed (filtered, averaged) values.
}
\description{
This method extracts raw or processed data from an \code{EPhysData} object.
}
\details{
The \code{GetData} function extracts the recorded data from an \code{EPhysData} or related object.
By default, the resulting data matrix contains unfiltered data from all repeated measurements.
Note, the main difference to the \link{as.data.frame}-Method is that this function returns a matrix in the wide format and without a time and stimulus trace.
}
\examples{
myEPhysData <- makeExampleEPhysData(sample_points = sample(seq(1, 400, 10), 1),replicate_count = sample(1:5, 1))
myEPhysData
# Get raw data
raw_data <- GetData(myEPhysData, Raw = TRUE)
head(raw_data)
# Get processed (filtered and averaged) data
AverageFunction(myEPhysData) <- median
filtered_and_averaged_data <- GetData(myEPhysData, Raw = FALSE)
head(filtered_and_averaged_data)
}
\seealso{
\code{\link{EPhysData-class}} \code{\link{TimeTrace}}, \code{\link{Rejected}}
}
