% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Set_EPhysData.R
\docType{methods}
\name{Get_Set_EPhysData}
\alias{Get_Set_EPhysData}
\alias{Rejected}
\alias{Rejected,EPhysData-method}
\alias{Rejected,EPhysData,ANY-method}
\alias{Rejected<-}
\alias{Rejected<-,EPhysData-method}
\alias{`Rejected<-`,EPhysData,ANY-method}
\alias{FilterFunction}
\alias{FilterFunction,EPhysData-method}
\alias{FilterFunction,EPhysData,ANY-method}
\alias{FilterFunction<-}
\alias{FilterFunction<-,EPhysData-method}
\alias{`FilterFunction<-`,EPhysData,ANY-method}
\alias{AverageFunction}
\alias{AverageFunction,EPhysData-method}
\alias{AverageFunction,EPhysData,ANY-method}
\alias{AverageFunction<-}
\alias{AverageFunction<-,EPhysData-method}
\alias{`AverageFunction<-`,EPhysData,ANY-method}
\alias{TimeTrace}
\alias{TimeTrace,EPhysData-method}
\alias{TimeTrace,EPhysData,ANY-method}
\alias{StimulusTrace}
\alias{StimulusTrace,EPhysData-method}
\alias{StimulusTrace,EPhysData,ANY-method}
\title{Get/Set methods for EPhysData objects}
\usage{
Rejected(X, return.fx = F, ...)

\S4method{Rejected}{EPhysData}(X, return.fx = F)

Rejected(X, ...) <- value

\S4method{Rejected}{EPhysData}(X) <- value

FilterFunction(X, ...)

\S4method{FilterFunction}{EPhysData}(X)

FilterFunction(X, ...) <- value

\S4method{FilterFunction}{EPhysData}(X) <- value

AverageFunction(X, ...)

\S4method{AverageFunction}{EPhysData}(X)

AverageFunction(X, ...) <- value

\S4method{AverageFunction}{EPhysData}(X) <- value

TimeTrace(X)

\S4method{TimeTrace}{EPhysData}(X)

StimulusTrace(X)

\S4method{StimulusTrace}{EPhysData}(X)
}
\arguments{
\item{X}{An \code{EPhysData} object}

\item{return.fx}{For \code{Rejected()}: Whether to return the function or the resulting logical vector. Default is \code{FALSE}, i.e. to return the function.}

\item{...}{Currently unused.}

\item{value}{A value (usually a function) to set.}
}
\description{
These methods are used to get and set the non-data slots from \link{EPhysData} objects.
}
\details{
\code{Rejected}: These functions set or get a function returning a logical vector indicating which of the repeated measurements stored in an \link{EPhysData} object to exclude from averaging. The following function from the EPhysMethods package may be helpful: \link[EPhysMethods:autoreject.by.distance]{EPhysMethods::autoreject.by.distance}, \link[EPhysMethods:autoreject.by.signalfree]{EPhysMethods::autoreject.by.signalfree}

\code{FilterFunction}: Set  a function for filtering each individual of the repeated measurements in the \link{EPhysData} object. Could be downsampling or noise removal, for instance. The following functions from the EPhysMethods package may be helpful: \link[EPhysMethods:filter.bandpass]{EPhysMethods::filter.bandpass}, \link[EPhysMethods:filter.detrend]{EPhysMethods::filter.detrend},

\code{AverageFunction}: Set a function describing how averaging across repeated measurement should be performed in the \link{EPhysData} object. Usually, \link[base:mean]{mean} can be a good start.

\code{TimeTrace}: These functions set or get the time trace belonging to the measurements stored in an \link{EPhysData} object.

\code{StimulusTrace}: These functions set or get the stimulus trace belonging to the measurements stored in an \link{EPhysData} object.
}
\examples{
# Create an EPhysData object with example data
myEPhysData <- makeExampleEPhysData()

# Get the "Rejected" slot
Rejected(myEPhysData)
head(GetData(myEPhysData))

# Set the "Rejected" slot
Rejected(myEPhysData) <- sample(c(TRUE,FALSE), dim(myEPhysData)[2], TRUE)
Rejected(myEPhysData)
head(GetData(myEPhysData))
head(GetData(myEPhysData,Trials=1:dim(myEPhysData)[2]))

# Get the "filter.fx" slot
FilterFunction(myEPhysData)

# Set the "filter.fx" slot
FilterFunction(myEPhysData) <- scale
FilterFunction(myEPhysData)

# Get the "average.fx" slot
AverageFunction(myEPhysData)
head(GetData(myEPhysData))

# Set the "average.fx" slot
AverageFunction(myEPhysData) <- median
AverageFunction(myEPhysData)
head(GetData(myEPhysData))
head(GetData(myEPhysData,Raw=T))

}
\seealso{
\link[EPhysMethods:autoreject.by.distance]{EPhysMethods::autoreject.by.distance}, \link[EPhysMethods:autoreject.by.signalfree]{EPhysMethods::autoreject.by.signalfree}, \link[EPhysMethods:filter.bandpass]{EPhysMethods::filter.bandpass}, \link[EPhysMethods:filter.detrend]{EPhysMethods::filter.detrend},
}
