% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEPhysData.R
\name{ggEPhysData-methods}
\alias{ggEPhysData-methods}
\title{Generate a ggplot2 plot for EPhysData objects}
\value{
A ggplot2 plot visualizing the EPhysData.
}
\description{
This method generates a ggplot2 plot for objects of class 'EPhysData'. It allows you to visualize electrical physiology data over time.
}
\examples{
ephys_data <- makeExampleEPhysData(,replicate_count = )  # Create an EPhysData object
Rejected(ephys_data) <- c(F,F,F,F,T,T,T,F,F,F,F,T,F,T,F) # randmly reject some
AverageFunction(ephys_data) <- mean
ggEPhysData(ephys_data)  # Generate a ggplot2 plot
}
