% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0EPhysSet-class.R
\name{newEPhysSet}
\alias{newEPhysSet}
\title{Create an instance of the EPhysSet class}
\usage{
newEPhysSet(Data, Metadata)
}
\arguments{
\item{Data}{A list of \linkS4class{EPhysData} objects.}

\item{Metadata}{A data frame containing metadata information associated with the data, each row corresponds to one list item.
\describe{
\item{Step}{A character vector containing the steps associated with the data.}
\item{Eye}{A character vector containing the possible values "RE" (right eye) and "LE" (left eye).}
\item{Channel}{A character vector containing the unique names of the third level of \code{Data}.}
}}
}
\value{
An object of class \code{EPhysSet}.
}
\description{
This function creates an instance of the \code{EPhysSet} class
with the specified data and attributes.
}
\examples{
# Create example data
# Create metadata and an EPhysSet object
metadata <- data.frame(StepID = c("A1", "A2", "A3", "A4"))
ephysSet <-
  newEPhysSet(
    Data = list(
      makeExampleEPhysData(),
      makeExampleEPhysData(),
      makeExampleEPhysData(),
      makeExampleEPhysData()
    ),
    Metadata = metadata
 )
ephysSet
}
\seealso{
\code{\link{EPhysSet-class}}
}
